refactor: 이동 전략이 기물 대신 행동 표현하게 변경
이동 전략 자체는 다른 기물을 고려하지 않으므로
차와 포의 이동이 동일하다.

하지만 기존의 기물을 표현하는 이동 전략(ChariotMovement 등)은
차와 포 각각의 이동 전략을 똑같은 코드의 두 클래스 파일로 구현해야 하는
비효율이 있었다.

이를 해결하기 위해 XxxMovement가 기물이 아닌 행동을 표현하도록 한다.
